feat(react): implement prepare outside React context (FEP-2357)
Extract the usePrepare body into a context-free `makePrepare` factory that
takes the three stackflow() inputs it actually depends on (config, loadData,
activityComponentMap) and returns the `Prepare` function. Wire it as the
`prepare` field of the stackflow() output so chunk/data preloading can run
before the React tree mounts, and refactor `usePrepare` into a thin wrapper
over the same implementation so in-tree callers are unchanged.
